chore: InputStateProvider 제거하고 props로 대체 	modified:   components/layouts/main/card/CardList.tsx 	modified:   components/layouts/main/searchBar/SearchBar.tsx 	deleted:    contexts/InputStateProvider.tsx 	modified:   pages/_app.tsx 	modified:   pages/folder/index.tsx
