+++ components/layouts/main/card/CardList.tsx
@@ -1,5 +1,4 @@
 import classNames from 'classnames';
-import { useContext } from 'react';
 
 import { LOADING_MESSAGE } from 'constants/constants';
 import useFetch from 'hooks/useFetch';
@@ -8,7 +7,6 @@ import { LINKS_API_URL, LINKS_FOLDER_ID_API_URL, LinksApiResponse } from '@/apis
 import ErrorMessage from '@/components/common/ErrorMessage';
 import Card from '@/components/layouts/main/card/Card';
 import styles from '@/components/layouts/main/card/CardList.module.css';
-import { InputStateContext } from '@/contexts/InputStateProvider';
 
 const cardListClasses = classNames(styles['card-list'], styles.grid, 'grid', 'width-full');
 const noCardListTextBoxClasses = classNames(
@@ -20,27 +18,26 @@ const noCardListTextBoxClasses = classNames(
 
 interface CardListProps {
   folderId?: number;
+  filter?: string;
 }
 
-function CardList({ folderId = 0 }: CardListProps) {
+function CardList({ folderId = 0, filter = '' }: CardListProps) {
   const url = folderId === 0 ? LINKS_API_URL : LINKS_FOLDER_ID_API_URL(folderId);
   const { data, loading, error } = useFetch<LinksApiResponse>(url);
-
-  const { inputState } = useContext(InputStateContext);
 
   // {created_at, description, folder_id, id, image_source, title, updated_at, url}
   const linkList = data?.data ?? [];
   const linkCount = linkList?.length ?? 0;
 
   // 입력 상태를 소문자로 변환
-  const inputStateLower = inputState.toLowerCase();
+  const filterLowerCase = filter.toLowerCase();
 
   // 필터링할 때 소문자로 변환하여 대소문자 구분 없이 비교
   const filteredLinkList = linkList.filter(
     (link) =>
-      link.title?.toLowerCase().includes(inputStateLower) ||
-      link.description?.toLowerCase().includes(inputStateLower) ||
-      link.url.toLowerCase().includes(inputStateLower)
+      link.title?.toLowerCase().includes(filterLowerCase) ||
+      link.description?.toLowerCase().includes(filterLowerCase) ||
+      link.url.toLowerCase().includes(filterLowerCase)
   );
 
   return (
